requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.8.1=pypi_0
attrs=23.2.0=pypi_0
bzip2=1.0.8=h6c40b1e_5
ca-certificates=2024.3.11=hecd8cb5_0
django=4.1.13=pypi_0
djangorestframework=3.15.1=pypi_0
dnspython=2.6.1=pypi_0
drf-spectacular=0.27.1=pypi_0
inflection=0.5.1=pypi_0
iniconfig=2.0.0=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
libffi=3.4.4=hecd8cb5_0
mysqlclient=2.2.4=pypi_0
ncurses=6.4=hcec6c5f_0
openssl=3.0.13=hca72f7f_0
packaging=24.0=pypi_0
pip=23.3.1=py311hecd8cb5_0
pluggy=1.4.0=pypi_0
pytest=8.1.1=pypi_0
pytest-django=4.8.0=pypi_0
pytest-mock=3.14.0=pypi_0
pytest-progress=1.2.5=pypi_0
pytest-sugar=1.0.0=pypi_0
python=3.11.8=hf27a42d_0
python-dotenv=1.0.1=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=hca72f7f_0
referencing=0.34.0=pypi_0
rpds-py=0.18.0=pypi_0
setuptools=68.2.2=py311hecd8cb5_0
sqlite=3.41.2=h6c40b1e_0
sqlparse=0.2.4=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.12=h5d9f67b_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pypi_0
wheel=0.41.2=py311hecd8cb5_0
xz=5.4.6=h6c40b1e_0
zlib=1.2.13=h4dc903c_0
